feat(config,wiring): #216 #217 — surface pool tunables + activation/retry/poll
  * Add MaxConns/MinConns/MaxConnLifetime/MaxConnIdleTime/HealthCheckPeriod
    to DatabaseConfig with documented production defaults
    (25/5/1h/30m/30s).
  * applySets cases for each, parsed via strconv.ParseInt(...,10,32) /
    time.ParseDuration.
  * buildStorageConfig forwards the values to pool.Config.
    {MaxConnections,MinConnections,MaxConnectionLifetime,IdleTimeout,
    HealthCheckInterval} so pool.Open applies them to pgxpool.Config.

  * Add cfg.Auth.ActivationTimeout (default 30s); wiring.go now reads
    it instead of the literal 30s for Deps.ActivationTimeout.
  * Add cfg.Pipeline.Processor.IdlePollInterval (default 200ms); the
    HL7 processor wiring honors hl7_processor.idle_poll_interval first
    then falls through to Pipeline.Processor.IdlePollInterval, keeping
    the existing 200ms safety floor.
  * Promote Pipeline.Scheduler from StageConfig to SchedulerConfig
    (StageConfig embedded inline so YAML keys at
    pipeline.scheduler.{claim_batch_size,idle_poll_interval,...}
    keep their existing positions); add SchedulerRetryConfig with
    Initial/Max/Min/MaxAttempts (defaults 1s/30s/500ms/8 — the
    previously-hardcoded values). Wiring layer reads from
    cfg.Pipeline.Scheduler.Retry.

Each new key is reachable through both YAML and --set; loadConfig
re-applies defaults post-decode so a YAML that omits a block keeps
production behavior. docs_lint_test.go's pathExistsInStruct now
recurses through embedded structs tagged `yaml:",inline"` so the
SchedulerConfig→StageConfig flattening doesn't trip the silent-
swallow guard.

+++ cmd/fhir-subs/config_storage_hygiene_test.go
@@ -176,7 +176,7 @@ pipeline:
 	if got, want := cfg.Pipeline.Scheduler.Retry.Min, 750*time.Millisecond; got != want {
 		t.Errorf("Pipeline.Scheduler.Retry.Min: got %v want %v", got, want)
 	}
-	if got, want := cfg.Pipeline.Scheduler.Retry.MaxAttempts, 12; got != want {
+	if got, want := cfg.Pipeline.Scheduler.Retry.MaxAttempts, int32(12); got != want {
 		t.Errorf("Pipeline.Scheduler.Retry.MaxAttempts: got %d want %d", got, want)
 	}
 }
@@ -226,7 +226,7 @@ func TestApplySets_StorageHygiene217_ActivationAndPipelineTunables(t *testing.T)
 			key:  "pipeline.scheduler.retry.max_attempts",
 			val:  "20",
 			got:  func(c *Config) any { return c.Pipeline.Scheduler.Retry.MaxAttempts },
-			want: 20,
+			want: int32(20),
 		},
 	}
 	for _, c := range cases {
@@ -286,7 +286,7 @@ func TestDefaultConfig_StorageHygiene217_SchedulerRetry(t *testing.T) {
 	if got, want := cfg.Pipeline.Scheduler.Retry.Min, 500*time.Millisecond; got != want {
 		t.Errorf("default Scheduler.Retry.Min: got %v want %v", got, want)
 	}
-	if got, want := cfg.Pipeline.Scheduler.Retry.MaxAttempts, 8; got != want {
+	if got, want := cfg.Pipeline.Scheduler.Retry.MaxAttempts, int32(8); got != want {
 		t.Errorf("default Scheduler.Retry.MaxAttempts: got %d want %d", got, want)
 	}
 }
